[XDLD-234] update button icon props
