--- packages/ui-components/src/components/button/mock.ts
@@ -1,15 +1,0 @@
-import { ButtonProps } from './button';
-
-export const mockButtonProps: Record<ButtonProps['type'], ButtonProps> = {
-  button: {
-    type: 'button',
-    children: 'button',
-    onClick: () => alert('hello world'),
-  },
-  link: {
-    type: 'link',
-    children: 'link',
-    href: 'https://github.com/',
-    title: 'Github',
-  },
-};
